python-ai-caller/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
twilio==8.11.0
langchain-community==0.0.20
langchain-openai==0.0.5
python-dotenv==1.0.0
openai==1.10.0
mysql-connector-python==8.3.0
pytz==2024.1
httpx==0.26.0
faiss-cpu==1.7.4
